Add support for AWSCredentialsProvider paradigm for S3 Backups
