+++ exhibitor-core/src/main/java/com/netflix/exhibitor/core/s3/S3ClientFactoryImpl.java
@@ -27,186 +27,13 @@ public class S3ClientFactoryImpl implements S3ClientFactory
     @Override
     public S3Client makeNewClient(final S3Credential credentials) throws Exception
     {
-        return new S3Client()
-        {
-            private final AtomicReference<RefCountedClient>   client = new AtomicReference<RefCountedClient>(null);
-            {
-                changeCredentials(credentials);
-            }
-
-            @Override
-            public void changeCredentials(S3Credential credential) throws Exception
-            {
-                RefCountedClient   newRefCountedClient = (credential != null) ? new RefCountedClient(new AmazonS3Client(new BasicAWSCredentials(credentials.getAccessKeyId(), credentials.getSecretAccessKey()))) : new RefCountedClient(new AmazonS3Client());
-                RefCountedClient   oldRefCountedClient = client.getAndSet(newRefCountedClient);
-                if ( oldRefCountedClient != null )
-                {
-                    oldRefCountedClient.markForDelete();
-                }
-            }
-
-            @Override
-            public void close() throws IOException
-            {
-                try
-                {
-                    changeCredentials(null);
-                }
-                catch ( Exception e )
-                {
-                    throw new IOException(e);
-                }
-            }
-
-            @Override
-            public InitiateMultipartUploadResult initiateMultipartUpload(InitiateMultipartUploadRequest request) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.initiateMultipartUpload(request);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public PutObjectResult putObject(PutObjectRequest request) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.putObject(request);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public S3Object getObject(String bucket, String key) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.getObject(bucket, key);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public ObjectMetadata getObjectMetadata(String bucket, String key) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.getObjectMetadata(bucket, key);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public ObjectListing listObjects(ListObjectsRequest request) throws Exception
-            {
-                RefCountedClient    holder = client.get();
-                AmazonS3Client      amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.listObjects(request);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public ObjectListing listNextBatchOfObjects(ObjectListing previousObjectListing) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.listNextBatchOfObjects(previousObjectListing);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public void deleteObject(String bucket, String key) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    amazonS3Client.deleteObject(bucket, key);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public UploadPartResult uploadPart(UploadPartRequest request) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    return amazonS3Client.uploadPart(request);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public void completeMultipartUpload(CompleteMultipartUploadRequest request) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    amazonS3Client.completeMultipartUpload(request);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-
-            @Override
-            public void abortMultipartUpload(AbortMultipartUploadRequest request) throws Exception
-            {
-                RefCountedClient holder = client.get();
-                AmazonS3Client amazonS3Client = holder.useClient();
-                try
-                {
-                    amazonS3Client.abortMultipartUpload(request);
-                }
-                finally
-                {
-                    holder.release();
-                }
-            }
-        };
+        return new S3ClientImpl(credentials);
     }
+
+    @Override
+    public S3Client makeNewClient(S3CredentialsProvider credentialsProvider) throws Exception
+    {
+        return new S3ClientImpl(credentialsProvider);
+    }
+
 }
